Reproducing ROC curves (#3)
* Remove duplicate keys from attack parameter dictionaries.

Fixes #1.

* Update default number of training epochs on cats and dogs dataset.

Fixes #2. According to the first author this value was used to produce
the results in the paper. Indeed, this configuration gives ROC curves
for the evaluation on the cats and dogs dataset, which are much more
similar to those presented in figure 14.

* Update training.

Actually call the defined function instead of just trying to load the
dataset file.

* Update number of images in evaluation set.

The number of images of adversarial and non-adversarial examples is
updated in order to make the ROC to look alike those of the paper's
figure 14 - in terms of smoothness.

* Add instructions to reproduce ROC plots.

Just some short guiding description on how to use the code to get ROC
curves that look very similar as those of the paper's figure 14.